Count a program's instructions in a long and keep Program.Current named
A program counted the instructions it had run in an int, which one that
stays up long enough would wrap into a negative number. The count is a
long now, through the language contract and the save.

Program.Current was made with the name the program had the first time
it was read, so a program that named itself afterwards kept reading its
old name there. SetName now renames that object as well.

+++ computers/src/main/java/dev/jstech/computers/machine/CannonProgram.java
@@ -74,7 +74,7 @@ final class CannonProgram implements ILanguageProcess {
     }
 
     @Override
-    public int spent() {
+    public long spent() {
         return this.process.spent();
     }
