+++ computers/src/main/java/dev/jstech/computers/vm/program/Snapshot.java
@@ -24,7 +24,7 @@ import java.util.Map;
  */
 public record Snapshot(long heapBudget, List<IHeld> held, List<ThreadShot> threads, List<FrameShot> waiting,
                        Map<String, Map<String, IValue>> statics, IValue script, List<WatchShot> watches,
-                       List<String> console, int written, String state, String message, int spent,
+                       List<String> console, int written, String state, String message, long spent,
                        String name, List<MonitorShot> monitors, int nextThread, List<String> args,
                        int machineId, boolean exited, int exitCode, IValue onMessage, List<IValue> windows,
                        long nextWindow, long nextWidget, IValue onGatewayMessage, String gateway) {
@@ -49,7 +49,7 @@ public record Snapshot(long heapBudget, List<IHeld> held, List<ThreadShot> threa
     public Snapshot(final long heapBudget, final List<IHeld> held, final List<ThreadShot> threads,
                     final List<FrameShot> waiting, final Map<String, Map<String, IValue>> statics,
                     final IValue script, final List<WatchShot> watches, final List<String> console,
-                    final int written, final String state, final String message, final int spent,
+                    final int written, final String state, final String message, final long spent,
                     final String name, final List<MonitorShot> monitors, final int nextThread,
                     final List<String> args, final int machineId, final boolean exited, final int exitCode,
                     final IValue onMessage) {
